chore: delete preferences slice
